requirements.txt
Flask==1.1.1
numpy==1.16.4
Pillow==6.1.0
uWSGI==2.0.17.1